[REFACTOR] jwt 중복 파싱 수정 #80
