+++ src/main/java/umc/puppymode/service/AuthService/AppleAuthCommandService.java
@@ -5,7 +5,6 @@ import umc.puppymode.web.dto.LoginResponseDTO;
 
 public interface AppleAuthCommandService {
     LoginResponseDTO loginWithApple(String authorizationCode, String identityToken, String username, String fcmToken);
-    boolean verifyIdentityToken(String identityToken);
     AppleTokenResponseDTO getAppleTokens(String authorizationCode);
     String generateClientSecret();
 }
