+++ src/main/java/umc/puppymode/service/AuthService/AppleAuthQueryServiceImpl.java
@@ -7,10 +7,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.jwt.JwtException;
 import org.springframework.stereotype.Service;
+import umc.puppymode.config.AppleAuthConfig;
 import umc.puppymode.domain.enums.AuthProvider;
 import umc.puppymode.web.dto.UserAuthInfoDTO;
 
 import java.security.PublicKey;
+import java.util.Date;
 
 @Slf4j
 @Service
@@ -18,6 +20,59 @@ import java.security.PublicKey;
 public class AppleAuthQueryServiceImpl implements AppleAuthQueryService {
 
     private final AppleKeyService appleKeyService;
+    private final AppleAuthConfig appleAuthConfig;
+    private static final String APPLE_ISSUER = "https://appleid.apple.com";
+
+    /**
+     * Identity Token 을 검증합니다.
+     *
+     * @param identityToken
+     * @return Claims
+     */
+    public Claims verifyIdentityToken(String identityToken) {
+        try {
+            // Identity Token JWT 헤더에서 kid 값 추출
+            String kid = extractKidFromToken(identityToken);
+
+            // 공개 키 조회
+            PublicKey publicKey = appleKeyService.getApplePublicKey(kid);
+            if (publicKey == null) {
+                throw new IllegalArgumentException("No matching public key found for kid: " + kid);
+            }
+
+            // Id Token JWT 검증 수행
+            Jws<Claims> claims = Jwts.parserBuilder()
+                    .setSigningKey(publicKey)
+                    .build()
+                    .parseClaimsJws(identityToken);
+
+            Claims body = claims.getBody();
+
+            // iss(발급자) 검증
+            if (!APPLE_ISSUER.equals(body.getIssuer())) {
+                throw new IllegalArgumentException("Invalid issuer: " + body.getIssuer());
+            }
+
+            // aud(Audience) 검증
+            if (!appleAuthConfig.getClientId().equals(body.getAudience())) {
+                throw new IllegalArgumentException("Invalid audience: " + body.getAudience());
+            }
+
+            // exp(만료 시간) 검증
+            if (body.getExpiration().before(new Date())) {
+                throw new IllegalArgumentException("Token has expired");
+            }
+
+            // sub(사용자 고유 ID) 반환
+            String appleUserId = body.getSubject();
+
+            return body;
+
+        } catch (Exception e) {
+            log.error("Identity Token 검증 실패: {}", e.getMessage());
+            throw new IllegalArgumentException("Invalid Identity Token");
+        }
+    }
 
     /**
      * Identity Token 에서 사용자 정보 추출
@@ -28,21 +83,16 @@ public class AppleAuthQueryServiceImpl implements AppleAuthQueryService {
     @Override
     public UserAuthInfoDTO getUserInfo(String identityToken) {
         try {
-            PublicKey publicKey = appleKeyService.getApplePublicKey(identityToken);
-            Jws<Claims> claims = Jwts.parserBuilder()
-                    .setSigningKey(publicKey)
-                    .build()
-                    .parseClaimsJws(identityToken);
+            Claims claims = verifyIdentityToken(identityToken);
 
-            Claims payload = claims.getBody();
-            String userAuthId = payload.get("sub", String.class);
-            String email = payload.get("email", String.class);
-            Boolean emailVerified = Boolean.parseBoolean(payload.get("email_verified", String.class));
+            String appleAuthId = claims.get("sub", String.class);
+            String email = claims.get("email", String.class);
+            Boolean emailVerified = Boolean.parseBoolean(claims.get("email_verified", String.class));
 
             log.info("Apple Identity Token 정보 추출 완료");
 
             return UserAuthInfoDTO.builder()
-                    .userAuthId(userAuthId)
+                    .userAuthId(appleAuthId)
                     .email(email)
                     .authProvider(AuthProvider.APPLE)
                     .build();
@@ -50,5 +100,17 @@ public class AppleAuthQueryServiceImpl implements AppleAuthQueryService {
             log.error("Invalid Apple Identity Token", e);
             throw new IllegalArgumentException("Invalid Apple Identity Token");
         }
+    }
+
+    /**
+     * Id Token JWT 헤더에서 kid 값을 추출합니다.
+     */
+    private String extractKidFromToken(String identityToken) {
+        return Jwts.parserBuilder()
+                .build()
+                .parseClaimsJws(identityToken)
+                .getHeader()
+                .get("kid")
+                .toString();
     }
 }
